data-frame-tool-clean/requirements.txt
pandas>=2.1,<3.0
openpyxl>=3.1,<4.0
pyyaml>=6.0,<7.0
pandera>=0.18,<0.20
# Optional SQL/analytics extras used by connectors and CLI integrations
sqlalchemy>=2.0,<3.0
pyodbc>=5.0,<6.0
duckdb>=1.1,<2.0
pyarrow>=14.0,<15.0
pytest>=7.4,<8.0
